+++ maptasker/src/guimap.py
@@ -100,12 +100,14 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
             # Only deal with lines that have a style and class, or a 'Style tag details'
             font = ""
             decor = ""
+            table = False
             if (line and "style=" in line and "class=" in line) or in_style:
                 try:
                     # Get the style details: color font, heading size, bold, italicised.
                     temp = line.replace("&nbsp;", " ").split('style="')
                     style = previous_style if in_style else temp[1].split(";")
 
+                    # Go thru each style and capture appropriate details
                     for specific_style in style:
                         if "color:" in specific_style:
                             color = specific_style.replace("color:", "")
@@ -126,6 +128,8 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
                         elif "font-weight:" in specific_style:
                             font_to_use = "bold" if "bold" in specific_style else "normal"
                             font = f"{font};bold" if font else font_to_use
+                        elif specific_style == "is_table":
+                            table = True
                         else:
                             # Drop here if there is a ';' in the label.  This is a problem
                             text = line.split('class="h6-text">')[1].replace("</span>", "")
@@ -181,6 +185,7 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
                     last_item["color"] == color
                     and last_item["highlights"] == font
                     and not ("<a href=" in text or "<a href=" in last_item["text"])
+                    and not table
                 ):
                     # If they match, concatenate the text in the processed_line_data list
                     last_item["text"] += text
@@ -198,6 +203,7 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
                         spacing,
                         lblend,
                         decor,
+                        table,
                     )
                     break
 
@@ -218,6 +224,7 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
                 spacing,
                 lblend,
                 decor,
+                table,
             )
 
         line_num += 1
@@ -245,6 +252,7 @@ def process_label_html(lines: list, output_lines: dict, line_num: int, spacing: 
                 "spacing": processed_line_data[0]["spacing"],
                 "end": [item["end"] for item in processed_line_data],
                 "decor": [item["decor"] for item in processed_line_data],
+                "table": [item["table"] for item in processed_line_data],
             }
             # Bump our line numbers to point to next line.
             processed_line_data = []
@@ -279,6 +287,7 @@ def add_line_data(
     spacing: int,
     lblend: bool,
     decor: str,
+    table: bool,
 ) -> None:
     r"""
     Splits a string by newline characters and appends a dictionary for each
@@ -298,6 +307,8 @@ def add_line_data(
         spacing (int): The spacing value for the text.
         lblend (bool): A flag indicating if this is the end of a label.
         decor (str): The text decoration style (e.g., "underline").
+        table (bool): A flag indicating if the text is part of a table.
+
     Return:
         The list with the line(s) added.
     """
@@ -314,6 +325,7 @@ def add_line_data(
                     "spacing": spacing,
                     "end": lblend,
                     "decor": decor,
+                    "table": table,
                 },
             )
         if not subtext: