+++ maptasker/src/userintr.py
@@ -82,7 +82,7 @@ from maptasker.src.maputil2 import save_window_position
 from maptasker.src.initparg import initialize_runtime_arguments
 from maptasker.src.lineout import LineOut
 from maptasker.src.mapai import valid_api_key
-from maptasker.src.mapit import clean_up_memory, mapit_all
+from maptasker.src.mapit import clean_up_memory, display_output, mapit_all
 from maptasker.src.maputils import (
     append_to_filename,
     clear_tasker_data,
@@ -244,10 +244,10 @@ class MyGui(customtkinter.CTk):
         # Reestabslish the window size since it might get changed by the deiconify call.
         self.geometry(window_position)
 
-        # FIX: For Development Only!
+        # CHG: For Development Only!
         # The following lines are for testing only.
         # self.event_handlers.diagram_event()
-        self.event_handlers.map_event()
+        # self.event_handlers.map_event()
         # self.event_handlers.ai_apikey_event()
         # self.event_handlers.upgrade_event()
 
@@ -3451,20 +3451,35 @@ class EventHandlers:
     # The 'ReRun' program button has been pressed.  Set the run flag and close the GUI
     def rerun_event(self) -> None:
         """
-        Resets the program state and exits.
+        Reload the GUI for a fresh start.
         Args:
             self: The class instance.
         Returns:
             None: Does not return anything.
-        - Sets the rerun flag to True to restart the program on next run
-        - Calls withdraw() to reset the program state
-        - Calls quit() twice to ensure program exits"""
+        """
         the_view = self.parent
-        # Reset the program state since it may have been previously set by the 'Map' view.
-        reset_primeitems_single_names()
-        the_view.rerun = True
-        the_view.textview.destroy()
-        the_view.cleanup_and_run(run_only=False)
+
+        # Remap everything with the current settings from the GUI.
+        the_view.remapit(clear_names=False)
+
+        # Setup to redisplay the output inm the browser.
+        # Get the output directory/folder path
+        my_output_dir = os.getcwd()
+        # Finally, write out all of the output that is queued up.
+        my_file_name = f"{PrimeItems.slash}MapTasker.html"
+        # These need to be off for the web browsere to display
+        PrimeItems.program_arguments["guiview"] = False
+        PrimeItems.program_arguments["ai_analyze"] = False
+        # Display the final results in the default web browser
+        display_output(my_output_dir, my_file_name)
+
+        reload_gui(the_view)
+
+        # # Reset the program state since it may have been previously set by the 'Map' view.
+        # reset_primeitems_single_names()
+        # the_view.rerun = True
+        # the_view.textview.destroy()
+        # the_view.cleanup_and_run(run_only=False)
 
     # The 'Exit' program button has been pressed.  Call it quits
     def exit_program_event(self) -> None: